handle more pce calls
we record a symbol table using recordz() but this requires symbol
references to be instantiated -- i have noticed a lot of free calls to
send() and new() which may require that we get the _GXXX from prolog.

or we can go through all existing source and reword the new() and send()
calls to use @ notation.

or we could just not maintain a database at all and see if we can get
away with it.